+++ src/components/DebugField.tsx
@@ -1,6 +1,6 @@
 import type { FormField } from "../FormField.ts";
 import { useFieldData } from "../hooks/useFieldData.ts";
-import { useBlurred } from "../hooks/useBlurred.ts";
+import { useIsBlurred } from "../hooks/useIsBlurred.ts";
 import { useFieldErrors } from "../hooks/useFieldErrors.ts";
 import { useIsChanged } from "../hooks/useIsChanged.ts";
 import type { DetailedHTMLProps, InputHTMLAttributes } from "react";
@@ -12,13 +12,13 @@ export type Props = {
 } & DefaultPreProps;
 export function DebugField({ field, ...rest }: Props) {
     const data = useFieldData(field);
+    const isBlurred = useIsBlurred(field);
     const isChanged = useIsChanged(field);
-    const blurred = useBlurred(field);
     const errors = useFieldErrors(field);
     const json = {
         path: field.toString(),
         data,
-        blurred,
+        isBlurred,
         isChanged,
         errors,
     }