+++ src/controls/Input.tsx
@@ -23,7 +23,7 @@ export function Input(props: Props) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         />