+++ src/controls/IntegerInput.tsx
@@ -27,7 +27,7 @@ export function IntegerInput(props: Props) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         />