+++ src/controls/NumberInput.tsx
@@ -21,7 +21,7 @@ export function NumberInput(props: Props) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         />