+++ src/controls/RadioButton.tsx
@@ -36,7 +36,7 @@ export function RadioButton<T>(props: Props<T>) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         />