+++ src/controls/Select.tsx
@@ -41,7 +41,7 @@ export function Select<T>(props: Props<T>) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         >