+++ src/controls/TextArea.tsx
@@ -19,7 +19,7 @@ export function TextArea(props: Props) {
                 onChange?.(e);
             }}
             onBlur={e => {
-                field.setBlurred(true);
+                field.setIsBlurred(true);
                 onBlur?.(e);
             }}
         />