+++ src/hooks/useIsBlurred.ts
@@ -1,0 +1,13 @@
+import { useCallback, useSyncExternalStore } from "react";
+import type { FormField } from "../FormField.ts";
+
+export function useIsBlurred(field: FormField<any>): boolean {
+    return useSyncExternalStore(
+        // Subscribe
+        useCallback(onStoreChange => field._internal.subscribeToIsBlurred(onStoreChange), [field]),
+        // Get snapshot
+        () => field.isBlurred(),
+        // Get server snapshot
+        () => field.isBlurred()
+    );
+}