+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Formula uses [semantic versioning](https://semver.org/).
 
 Formula is currently pre-v1. It's relatively stable, but I expect there will be some breaking changes before v1.
 
+# Unreleased
+
+- BREAKING CHANGE: useBlurred renamed to useIsBlurred
+- feat: useIsChanged, `<IsChanged>`
+- feat: `<IsBlurred>`
+
 # 0.0.7
 
 - feat: export NumberInput